+++ src/core/wee-string.h
@@ -56,7 +56,7 @@ extern char *string_iconv_to_internal (const char *charset, const char *string);
 extern char *string_iconv_from_internal (const char *charset,
                                          const char *string);
 extern int string_iconv_fprintf (FILE *file, const char *data, ...);
-extern char *string_format_size (unsigned long size);
+extern char *string_format_size (unsigned long long size);
 extern void string_encode_base64 (const char *from, int length, char *to);
 extern int string_decode_base64 (const char *from, char *to);
 extern int string_is_command_char (const char *string);
